Added basic manual lenses for Mattermost types
This doesn't incur any extra deps and doesn't require any kind of
template Haskell, but does involve a lot of boilerplate. This was
mostly generated via editor macro, but work is needed to keep it
up-to-date, so it might be a net win to switch to a TH-based
solution later.