Add InitialLoad and mmGetInitialLoad for requesting initial session info
The web client logs in without specifying a team name as part of the
login request. It then gets a constrained list of teams for the user
by requesting the "initial_load" endpoint. We do that here, now, with
the aforementioned type and request function. We don't use everything
that comes back in an initial_load response yet, so I don't think it's
prudent to stub out the fields until we understand and need them.

This change also removes the team name field from the Login request.

This change is a step on the path to making it possible to give the user
a choice of team at startup to match the UX provided by the web client.