Interpret empty ID fields as missing
Many ID fields returned by the server use an empty ID string to indicate
"no ID." Before this change we would consume these empty ID strings as
actual IDs - and then code using this library would erroneously think it
had real IDs to work with. This change makes it so that in fields where
we think it's possible to get empty IDs, we interpret those as missing
by using Maybe. This makes it possible for user code to deal more
appropriately with missing data.