websockets: don't re-throw the timeout exception to the thread that generated it
This change makes it so that the websocket timeout watchdog doesn't have
to handle the exception it throws. Since it's the only source of the
exception, we know when it is generated, so the flow control of the
throw site can just halt execution instead of using `forever` and
relying on an exception to kill the thread.