Move HTTP.Stream instance to newtype in different module
This creates a type MMConn, which wraps the underlying Connection for
the purposes of our HTTP library, and also puts it into the common
Util module. Before, we had to import Network.Mattermost for that
instance, which was less than ideal. This also moves much of the HTTP
helper functions from Network.Mattermost.Endpoints into a new,
non-exposed module, Network.Mattermost.Connection.