MMConn: honor "Connection: close" response headers
This change makes it so that responses handled in doRequest and
doUnauthRequest will cause us to close our own connection if we know the
server is doing the same, as indicated by the presence of a "Connection:
close" response header.

This means that when we use a connection from the connection pool, we
need to put it back in the pool only if we *didn't* disconnect, so we
have to track whether the connection has been explicitly closed as a
result of a response header.

To do that, we now have an IORef in the MMConn structure that we use to
book-keep whether we have closed the connection. This way we don't have
to try to use a connection and get an exception in order to find out
that it's closed, if it's a connection we closed ourselves. In addition,
we have to use a slightly lower-level API from the resource-pool package
to get control over how connections are returned to the connection pool
in order to make this work.

This change does *not* improve a preexisting/reasonable problem where a
connection closed without our knowledge would still raise an exception
when used.