submitRequest: on EOF exception, try the request exactly once more
This change is made due to an observation that the connection pool may
end up containing stale/severed connections. When we obtain such a
connection from the pool, it will already be poised to raise an EOF
exception on its next use. That also means it's likely there are other
connections in the pool that are also severed, waiting to cause
exceptions. So when we encounter an EOF exception we 1) destroy all idle
connections in the pool on this assumption and 2) try the original
request one more time. Doing it this way means that the second attempt
will cause the pool to create a fresh connection for us. In this case
further exceptions likely indicate a deeper problem with the network
environment, so those exceptions are just propagated.