Fix support for parsing websocket action responses
Websocket actions, such as "user typing" notifications, generate server
responses that we previously could not parse. We weren't parsing them
because the parser for websocket events wasn't aware of an alternative
response message structure. This patch adds a new type for action
responses. It adds a new type rather than adding a new constructor
to the event type because I think that's a less ugly API (the API in
this patch is not ideal). This patch also adds logic to the websocket
response parser to first attempt to parse incoming messages as websocket
events (the most common case) and then fall back to attempting a parse
as a websocket action response. If both fail, the message and the parse
exceptions are all logged.