Raise RateLimitExceptions on HTTP 429 API responses
This change adds a new exception type, RateLimitException, and raises it
when an API request of any type results in an HTTP 429 server response.
The exception provides fields for the three rate-limiting parameters
that we expect Mattermost servers to return. The new exception can be
raised by any version 4 API function in the package. It is the caller's
responsibility to handle the exception and determine whether to throttle
future API requests to avoid rate limit violations.