Add support for basic auth stacked upon other auth
Internal company servers may be additional secured by a basic
authentication username and password. This commit add the capability to
add basic auth credentials to the requests. The most interesting case is
BasicAuth + Bearer token. Both have to go into the Authorization header.
There are some rumors that comma separating them would be enough on the
internet but this is not working in my tests. As a work around the
bearer token is sent as a cookie which means we also need the csrf token
for post requests.