Post: parse and carry 'metadata' that is included in API messages
This change updates the Post type to parse the 'metadata' field that it
turns out is included in API messages. We didn't know about this, and we
don't know when it was added to the protocol, but using this would allow
clients to avoid making additional requests for per-post data. This
change parses only the file (attachment) list and reaction list for each
post; the other metadata fields are not parsed since it isn't clear that
they'll be useful yet.